requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pika==1.0.1
pkg-resources==0.0.0
python-dotenv==0.10.5
requests==2.22.0
urllib3==1.25.8
python-socketio==4.4.0
tauros-api==1.5.2
